requirements.txt
# Core Deep Learning Libraries
tensorflow==2.15.0
keras==2.15.0

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Scientific Computing
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
scipy==1.11.4

# Grad-CAM and Visualization
tf-keras-vis==0.8.5
pillow==10.1.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# Data Augmentation
albumentations==1.3.1

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.27.1

# Model Analysis
tensorboard==2.15.1

# Image Processing
imageio==2.33.0
scikit-image==0.22.0

# Optional: For deployment
gunicorn==21.2.0

# Optional: For DICOM support (medical imaging)
pydicom==2.4.4
